@@ -3,11 +3,9 @@
  */
 import React from 'react';
 var utils = require('./utils');
-var classNames = require('classnames');
 import ComposedComponent from './ComposedComponent';
-import DatePicker from 'material-ui/DatePicker/DatePicker';
-import IconButton from 'material-ui/IconButton';
-import Clear from 'material-ui/svg-icons/content/clear';
+import {IconButton, DatePicker, TextField} from '@material-ui/core';
+//import Clear from '@material-ui/core/SvgIcon';
 
 /**
  * There is no default number picker as part of Material-UI.
@@ -21,8 +19,8 @@ class Date extends React.Component {
     }
 
 
-    onDatePicked(empty, date) {
-        this.props.onChangeValidate(date);
+    onDatePicked(e) {
+        this.props.onChangeValidate(e.target.value);
     }
 
     render() {
@@ -32,25 +30,14 @@ class Date extends React.Component {
         }
 
         return (
-            <div style={{width: '100%', display: 'block'}} className={this.props.form.htmlClass}>
-                <DatePicker
-                    mode={'landscape'}
-                    autoOk
-                    floatingLabelText={this.props.form.title}
-                    hintText={this.props.form.title}
+            <div style={{width: '100%', display: 'block'}}>
+                <TextField
+                    id="date"
+                    label="Birthday"
+                    type="date"
+                    defaultValue="2017-05-24"
                     onChange={this.onDatePicked}
-                    onShow={null}
-                    onDismiss={null}
-                    value={value}
-                    disabled={this.props.form.readonly}
-                    style={this.props.form.style || {width: '90%', display: 'inline-block'}}/>
-                {this.props.value &&
-                    <IconButton ref="button"
-                        onClick={() => this.props.onChangeValidate('')}
-                        style={{position: 'relative', display: 'inline-block', top: '6px',right: '4px', padding: '0', width: '24px', height: '24px'}}>
-                        <Clear />
-                    </IconButton>
-                }
+                />
             </div>
         );
     }
