@@ -3,18 +3,22 @@ import utils from './utils';
 import classNames from 'classnames';
 import ComposedComponent from './ComposedComponent';
 
-import RadioButton from 'material-ui/RadioButton';
-import RadioButtonGroup from 'material-ui/RadioButton/RadioButtonGroup';
+import {Radio, 
+        RadioGroup,
+        FormControlLabel } from '@material-ui/core';
 
 class Radios extends React.Component {
 
     render() {
+        //console.log("VALUE", this.state.value);
         let items = this.props.form.titleMap.map(function(item, index) {
             return (
-                <RadioButton label={item.name}
-                             value={item.value}
-                             key={index}
-                             disabled={this.props.form.readonly}
+                <FormControlLabel
+                    label={item.name}
+                    value={item.value}
+                    disabled={this.props.form.readonly}
+                    control={<Radio
+                        checked={this.props.value === item.value} />}
                     />
             )
         }.bind(this));
@@ -22,9 +26,9 @@ class Radios extends React.Component {
         return (
             <span className={this.props.form.htmlClass}>
               <label className="control-lable">{this.props.form.title}</label>
-              <RadioButtonGroup defaultSelected={this.props.value} name={this.props.form.title} onChange={this.props.onChangeValidate}>
+              <RadioGroup name={this.props.form.title} onChange={(e) => this.props.onChangeValidate(e)}>
                   {items}
-              </RadioButtonGroup>
+              </RadioGroup>
             </span>
         );
     }
