@@ -3,7 +3,7 @@
  */
 import React from 'react';
 import ComposedComponent from './ComposedComponent';
-import TextField from 'material-ui/TextField';
+import {TextField} from '@material-ui/core';
 
 class TextArea extends React.Component {
 
@@ -11,13 +11,12 @@ class TextArea extends React.Component {
         // FIXME: Obviously fix rowsMax eventually..
         //console.log('TextArea', this.props.form);
         return (
-            <div className={this.props.form.htmlClass}>
                 <TextField
                     type={this.props.form.type}
-                    floatingLabelText={this.props.form.title}
-                    hintText={this.props.form.placeholder}
+                    label={this.props.form.title}
+                    helperText={this.props.form.placeholder}
                     onChange={this.props.onChangeValidate}
-                    errorText={this.props.error || this.props.errorText}
+                    error={this.props.error || this.props.errorText}
                     defaultValue={this.props.value}
                     multiLine
                     rows={this.props.form.rows}
@@ -25,7 +24,6 @@ class TextArea extends React.Component {
                     disabled={this.props.form.readonly}
                     style={this.props.form.style || {width: '100%'}}
                 />
-            </div>
         );
     }
 }
